Update Think Python to third edition
These are the changes in the new edition, according to the book webpage: "Each chapter of the book is a Jupyter notebook, so you can read the text, run the code, and work on the exercises, all in one place. [...] The new edition covers several new topics, including regular expressions and automated testing with doctest and unittest. And it includes more and better exercises."